+++ src/main/java/team/awesome/pet/service/PetRescueOrganizationService.java
@@ -1,5 +1,6 @@
 package team.awesome.pet.service;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -100,7 +101,7 @@ public class PetRescueOrganizationService implements PetRescueOrganizationDao {
   }
 
   @Override
-  public void removeShelterById(int id) {
+  public void removeShelterById(BigInteger id) {
     this.shelterDao.removeShelterById(id);
   }
