+++ src/main/java/team/awesome/pet/service/ShelterService.java
@@ -1,5 +1,6 @@
 package team.awesome.pet.service;
 
+import java.math.BigInteger;
 import java.util.Collection;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -31,7 +32,7 @@ public class ShelterService {
     this.shelterDao.updateShelter(shelter);
   }
 
-  public void removeShelterById(int id) {
+  public void removeShelterById(BigInteger id) {
     this.shelterDao.removeShelterById(id);
   }
 }
